backend/requirements/base.txt
# backend/requirements/base.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dateutil==2.8.2
pydantic==2.5.0
pydantic-settings==2.1.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
email-validator==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
pillow==10.1.0
bcrypt==4.1.1
cryptography==41.0.7
docker==6.1.3
pyyaml==6.0.1
psutil==5.9.6
httpx==0.25.2
websockets==12.0